third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/connect-src-anchor-ping-expected.txt

CONSOLE ERROR: Refused to connect to 'https://localhost:8443/resources/dummy.txt' because it violates the following Content Security Policy directive: "connect-src http://127.0.0.1:8000".
